Changed to uint32_t

+++ kernels/volk/volk_32f_stddev_and_mean_32f_x2.h
@@ -105,7 +105,7 @@ volk_32f_stddev_and_mean_32f_x2_generic(float* stddev, float* mean,
 #endif /* LV_HAVE_GENERIC */
 
 static inline float
-update_square_sum_1_val(const float SquareSum, const float Sum, const size_t len, const float val) {
+update_square_sum_1_val(const float SquareSum, const float Sum, const uint32_t len, const float val) {
   // Updates a sum of squares calculated over len values with the value val
   float n = (float) len;
   return SquareSum + 1.f/( n * (n + 1.f) ) * ( n*val - Sum ) * ( n*val - Sum );
@@ -113,7 +113,7 @@ update_square_sum_1_val(const float SquareSum, const float Sum, const size_t len
 
 static inline float
 add_square_sums(const float SquareSum0, const float Sum0, 
-                const float SquareSum1, const float Sum1, const size_t len) {
+                const float SquareSum1, const float Sum1, const uint32_t len) {
   // Add two sums of squares calculated over the same number of values, len
   float n = (float) len;
   return SquareSum0 + SquareSum1 + .5f / n * ( Sum0 - Sum1 )*( Sum0 - Sum1 );
@@ -121,7 +121,7 @@ add_square_sums(const float SquareSum0, const float Sum0,
 
 static inline void
 accrue_result( float* PartialSquareSums, float* PartialSums, 
-               const size_t NumberOfPartitions, const size_t PartitionLen) {
+               const uint32_t NumberOfPartitions, const uint32_t PartitionLen) {
   // Add all partial sums and square sums into the first element of the arrays
   uint32_t accumulators = NumberOfPartitions;
   uint32_t stages = 0;
@@ -133,7 +133,7 @@ accrue_result( float* PartialSquareSums, float* PartialSums,
 
   for (uint32_t s = 0; s < stages; s++ ) {
     accumulators /= 2;
-    size_t idx = 0;
+    uint32_t idx = 0;
     for (uint32_t a = 0; a < accumulators; a++)  {
       PartialSquareSums[idx] = add_square_sums(PartialSquareSums[idx], PartialSums[idx], 
                                PartialSquareSums[idx + offset], PartialSums[idx + offset],
